Part 1: I just made new integers for third, fourth, and fifth value then add them into the equation.
Part 2:For this I made number an integer in line 13 and added another parentheses in line 14.

Part 3: I made number1 an integer in line 10. Parentheses in line 13 and quotes in 14 to print text "Seven times your second number is." Also switched the 7( for 7 *

Part 4: I just remembered what to do from the activity in class and just mimicked that.

Part 5:first_name is integer that says What is your first name:. last_name is input asking for last name. then i printed a line saying "hello, first name+ ' ' for a space +last name.